backend/requirements.txt
asgiref==3.6.0
Django==3.1.14
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djoser==2.0.1
psycopg2-binary==2.8.6
pytz==2022.7
sqlparse==0.4.3
